- added fallbacks.

+++ client/src/constants/helper.ts
@@ -33,28 +33,31 @@ export const afterModalformFinish = (actionRef: MutableRefObject<ActionType>, ms
 }
 
 export const clearUserData = () => {
-	const cookies = document.cookie.split(';')
+	if (typeof document !== 'undefined') {
+		const cookies = document.cookie.split(';')
 
-	for (let i = 0; i < cookies.length; i++) {
-		const cookie = cookies[i]
-		const eqPos = cookie.indexOf('=')
-		const name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie
-		document.cookie = name + '=;expires=Thu, 01 Jan 1970 00:00:00 GMT'
+		for (let i = 0; i < cookies.length; i++) {
+			const cookie = cookies[i]
+			const eqPos = cookie.indexOf('=')
+			const name = eqPos > -1 ? cookie.substr(0, eqPos) : cookie
+			document.cookie = name + '=;expires=Thu, 01 Jan 1970 00:00:00 GMT'
+		}
+
+		caches.keys().then(keys => {
+			keys.forEach(key => caches.delete(key))
+		})
+
+		indexedDB.databases().then(dbs => {
+			dbs.forEach(db => indexedDB.deleteDatabase(db.name))
+		})
+
+		localStorage.clear()
+		sessionStorage.clear()
+
+		setTimeout(() => {
+			window.location.assign('/')
+		}, 1500)
 	}
-
-	caches.keys().then(keys => {
-		keys.forEach(key => caches.delete(key))
-	})
-	indexedDB.databases().then(dbs => {
-		dbs.forEach(db => indexedDB.deleteDatabase(db.name))
-	})
-
-	localStorage.clear()
-	sessionStorage.clear()
-
-	setTimeout(() => {
-		window.location.assign('/')
-	}, 1500)
 }
 
 export const capFrstLtr = (str: string) => str?.charAt(0).toUpperCase() + str?.slice(1)
